# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReloadCars API",
    "description": "API do app de pontos de recarga para veículos elétricos.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/charge-points/nearby": {
      "get": {
        "operationId": "getNearbyChargePoints",
        "summary": "Pontos de recarga próximos, do mais perto ao mais longe (RF08)",
        "tags": [
          "charge-points"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "lat",
            "required": true
          },
          {
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "lng",
            "required": true
          },
          {
            "schema": {
              "default": 10,
              "type": "number",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 50
            },
            "in": "query",
            "name": "radiusKm",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "chargePoints": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string",
                            "enum": [
                              "partner",
                              "ocm"
                            ]
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "address": {
                            "nullable": true,
                            "type": "string"
                          },
                          "latitude": {
                            "type": "number"
                          },
                          "longitude": {
                            "type": "number"
                          },
                          "powerKw": {
                            "nullable": true,
                            "type": "number"
                          },
                          "pricePerKwhCents": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "connectors": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "openingHours": {
                            "nullable": true,
                            "type": "string"
                          },
                          "attribution": {
                            "nullable": true,
                            "type": "string"
                          },
                          "partnerName": {
                            "nullable": true,
                            "type": "string"
                          },
                          "distanceMeters": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "source",
                          "name",
                          "description",
                          "address",
                          "latitude",
                          "longitude",
                          "powerKw",
                          "pricePerKwhCents",
                          "connectors",
                          "openingHours",
                          "attribution",
                          "partnerName",
                          "distanceMeters"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "chargePoints"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/charge-points/{id}": {
      "get": {
        "operationId": "getChargePoint",
        "summary": "Detalhes de um ponto de recarga",
        "tags": [
          "charge-points"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "partner",
                        "ocm"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number"
                    },
                    "longitude": {
                      "type": "number"
                    },
                    "powerKw": {
                      "nullable": true,
                      "type": "number"
                    },
                    "pricePerKwhCents": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "connectors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "openingHours": {
                      "nullable": true,
                      "type": "string"
                    },
                    "attribution": {
                      "nullable": true,
                      "type": "string"
                    },
                    "partnerName": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "source",
                    "name",
                    "description",
                    "address",
                    "latitude",
                    "longitude",
                    "powerKw",
                    "pricePerKwhCents",
                    "connectors",
                    "openingHours",
                    "attribution",
                    "partnerName"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/places/suggestions": {
      "get": {
        "operationId": "getPlaceSuggestions",
        "summary": "Sugestões de destino enquanto o usuário digita (RF05)",
        "tags": [
          "places"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 200
            },
            "in": "query",
            "name": "q",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "query",
            "name": "sessionToken",
            "required": true
          },
          {
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "lat",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "lng",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "address": {
                            "nullable": true,
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "address",
                          "type"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "suggestions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/places/{id}": {
      "get": {
        "operationId": "getPlace",
        "summary": "Coordenadas do destino escolhido",
        "tags": [
          "places"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "query",
            "name": "sessionToken",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "address": {
                      "nullable": true,
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number"
                    },
                    "longitude": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "address",
                    "latitude",
                    "longitude"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}